Refactor Kokoro TTS isolated service to use standalone worker script
- Extract embedded worker code into standalone kokoro_worker.py
- Simplify kokoro_tts_isolated.py by removing dynamic script creation
- Replace temporary file approach with version-controlled worker script
- Update bot.py to use refactored isolated service with proper imports
- Maintain same process isolation and Metal threading safety
- Improve maintainability and testability of worker code
