+++ server/bot.py
@@ -22,7 +22,8 @@ from pipecat.pipeline.runner import PipelineRunner
 from pipecat.pipeline.task import PipelineParams, PipelineTask
 from pipecat.processors.aggregators.openai_llm_context import OpenAILLMContext
 from pipecat.services.openai.llm import OpenAILLMService
-from kokoro_tts import KokoroTTSService
+# from kokoro_tts import KokoroTTSService
+from kokoro_tts_isolated import KokoroTTSIsolated
 from pipecat.services.whisper.stt import WhisperSTTServiceMLX, MLXModel
 from pipecat.transports.base_transport import TransportParams
 from pipecat.processors.frameworks.rtvi import RTVIConfig, RTVIObserver, RTVIProcessor
@@ -44,10 +45,6 @@ ice_servers = [
 
 SYSTEM_INSTRUCTION = """
 "You are Pipecat, a friendly, helpful chatbot.
-
-You are running a voice AI tech stack entirely locally, on macOS. Whisper for speech-to-text, a Qwen3 model with 235 billion parameters for language understanding, and Kokoro for speech synthesis. The pipeline also uses Silero VAD and the open source, native audio smart-turn v2 model.
-
-Your goal is to demonstrate your capabilities in a succinct way.
 
 Your input is text transcribed in realtime from the user's voice. There may be transcription errors. Adjust your responses automatically to account for these errors.
 
@@ -75,16 +72,15 @@ async def run_bot(webrtc_connection):
 
     stt = WhisperSTTServiceMLX(model=MLXModel.LARGE_V3_TURBO_Q4)
 
-    tts = KokoroTTSService(model="prince-canuma/Kokoro-82M", voice="af_heart", sample_rate=24000)
-    # Alternative: Process-isolated version to avoid Metal assertion failures on some systems
-    # from kokoro_tts_isolated import KokoroTTSIsolated
-    # tts = KokoroTTSIsolated(model="prince-canuma/Kokoro-82M", voice="af_heart", sample_rate=24000)
+    # tts = KokoroTTSService(model="prince-canuma/Kokoro-82M", voice="af_heart", sample_rate=24000)
+    # Process-isolated version to avoid Metal assertion failures (now refactored to use standalone worker)
+    tts = KokoroTTSIsolated(model="prince-canuma/Kokoro-82M", voice="af_heart", sample_rate=24000)
 
     llm = OpenAILLMService(
-        api_key=None,
+        api_key="dummyKey",
         model="google/gemma-3-12b",  # Medium-sized model. Uses ~8.5GB of RAM.
         # model="mlx-community/Qwen3-235B-A22B-Instruct-2507-3bit-DWQ", # Large model. Uses ~110GB of RAM!
-        base_url="http://localhost:8000/v1",
+        base_url="http://127.0.0.1:1234/v1",
         max_tokens=4096,
     )
